// webapp/src/@generated/schemas/monthlyReportDto.ts
/**
 * Generated by orval v6.12.1 🍺
 * Do not edit manually.
 * Vol
 * Registration api
 * OpenAPI spec version: 0.1
 */
import type { DailyVisitorDto } from './dailyVisitorDto';
import type { MonthlyAgeGroupDto } from './monthlyAgeGroupDto';

export interface MonthlyReportDto {
  total: number;
  visitors: DailyVisitorDto[];
  groups: MonthlyAgeGroupDto[];
}